Move authentication tag of chunks to the metadata object
This makes sure that even if the user knows the correct encryption key
to the chunk, they cannot rewrite the contents without also modifying
the metadata.

It should be impossible to anyone learning the `ChunkPointer` of a
chunk to rewrite the contents without detection unless they also have
the ability to make the changes in the metadata objects.

This change will make future asymmetric signing of metadata more robust.