In the benchmark, fix measurement of chunk reuse
The benchmark now collects all the distinct uses of all chunks, and
divides the number of uses (`creuse_sum`) with the number of
chunks (`creuse_cnt`).

If this fraction is >1, then a single chunk is used more than once in
the stash. Currently in my test set every chunk is used on average 1.3
times.