Fix Bug: too large of a record will overfull the object buffer
This is less of a fix and more of a mitigation. Here I just assume
that no single record, compressed, is larger than the object size.

The problem with this approach is that if a record is too big then the
crash happens anyway. Currently, this should happen when a file has
over 47000 chunks, or with an average chunk size of 64k, larger than 3TB.

I have a plan for several long-term fixes to this issue, but they
would take a while to implement, and this seems like a livable
trade-off at the moment.