Remove libsodium as dependency, replace blake2b with blake2b_simd
This moves all functionality that depended on libsodium to be based on
the `blake2b_simd` crate. I re-implemented the key derivation to be
mostly compatible with the `libsodium` scheme, except it's now using the
context string as a key to Blake2sb-HMAC.

The `rust-argon2` crate also depends on `blake2b_simd`, so there's no
redundant dependency for the same function.

As a side-effect of this change, and the tuning of Blake2sb, I have
now measured 1GB/s on 5 threads. Here:

stats for path (/Users/user/Code/rust/zerostash), seconds: 6.5285801580000005
 * files: 45140,
 * chunks: 70749,
 * data size: 6878.854786872864
 * throughput: 1002.7371203461443
 * objects: 1068
 * output size: 4272
 * compression ratio: 0.6210336069533541
 * meta dump time: 0.33149777
 * meta object count: 7
 * chunk reuse: 96054/70749 = 1.3576728999703176

read time: 1.149141747
restore time: 11.850139612
throughput packed: 328.63355150339123
throughput unpacked: 529.1719285782145